requirements.txt
emoji>=2.14.0
lingua-language-detector>=1.4.0
